+++ 2021/22_reactor_reboot/range.hh
@@ -8,7 +8,7 @@ struct Range
   int min;
   int max;
 
-  unsigned int size() const{return 1+max-min;}
+  long long  size() const{return 1+max-min;}
   set<Range> splitBy(Range const & tr) const;
   bool contains(int c) const {return (c>=min)and(c<=max);}
   int centre() const { return (min+max)/2; }
